requirements-linux.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
python-multipart
gunicorn
sqlalchemy
alembic
psycopg2-binary
geoalchemy2
passlib[argon2]
geoalchemy2[shapely]
minio
requests